requirements.txt
Django==1.9
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
Pillow==3.3.1
psycopg2==2.6.2
PyJWT==1.4.2
Unipath==1.1
